feat(FN-995): change format of time in email, move tests, send email to confirmation page
