+++ portal/templates/utilisation-report-service/utilisation-report-upload/confirmation.njk
@@ -10,7 +10,7 @@
                     {{ reportPeriod }} GEF report sent to UKEF
                 </h1>
             </div>
-            <p class="govuk-body" data-cy="paragraph">A confirmation email has been sent to {{ bankEmail }}.</p>
+            <p class="govuk-body" data-cy="paragraph">A confirmation email has been sent to {{ paymentOfficerEmail }}.</p>
             <a href="/logout" role="button" class="govuk-button" data-module="govuk-button" data-cy="logout-button">Sign out</a>
         </div>
     </div>
